// src/app/transport-service/booking/dto/booking-batch.dto.ts
export class BookingBatchDTO {
    bookingBatchID: string;
    batchCode: string;
    name: string;
    description: string;
    creationDate: Date;
    bookingSlipQuantity: number;
    customerPartyID: string;
    customer: string;
    bookingPartyID: string;
    bookingParty: string;
    statusTypeID: string;
    status: string;
}